feat(1737): Phase 2 - Validated page.
