+++ app/src/scenes/phase2/index.jsx
@@ -1,12 +1,10 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import { Switch, useHistory } from "react-router-dom";
-import { permissionPhase2, YOUNG_STATUS_PHASE2 } from "../../utils";
+import { permissionPhase2 } from "../../utils";
 import Mig from "./Mig";
 import Home from "./Home";
 import EditEquivalence from "./views/EditEquivalence";
-import ValidatedDesktop from "./desktop/Validated";
-import ValidatedMobile from "./mobile/Validated";
 import { SentryRoute } from "../../sentry";
 import useDocumentTitle from "../../hooks/useDocumentTitle";
 
@@ -17,18 +15,6 @@ export default function Index() {
   const history = useHistory();
 
   if (!young || !permissionPhase2(young)) history.push("/");
-
-  if (young.statusPhase2 === YOUNG_STATUS_PHASE2.VALIDATED)
-    return (
-      <>
-        <div className="hidden flex-1 md:flex">
-          <ValidatedDesktop />
-        </div>
-        <div className="flex md:hidden ">
-          <ValidatedMobile />
-        </div>
-      </>
-    );
 
   return (
     <Switch>
